+++ febr.c
@@ -12,7 +12,7 @@
 
 #include <stdio.h>
 #define LOG(fmt, ...) \
-	fprintf(stdout, "[%lx %s %d]"fmt, pthread_self(), __FUNCTION__, __LINE__,  ##__VA_ARGS__);
+	fprintf(stdout, "[%lx %s %d]"fmt, (unsigned long)pthread_self(), __FUNCTION__, __LINE__,  ##__VA_ARGS__);
 
 #define DBG(...) LOG(__VA_ARGS__)
